requirements.txt
ably==0.8.1
dj-database-url==0.4.1
Django==1.9.7
django-braces==1.9.0
django-cors-headers==1.1.0
django-mailgun==0.9.1
django-oauth-toolkit==0.10.0
django-reversion==2.0.4
djangorestframework==3.3.3
gunicorn==19.6.0
msgpack-python==0.4.8
oauthlib==1.0.3
pycrypto==2.6.1
requests==2.7.0
shortuuid==0.4.3
six==1.10.0
whitenoise==2.0.6
